altera assinatura das funções para a variavel tamanho

+++ include/ordenacao.h
@@ -1,11 +1,11 @@
 #pragma once
 
-void bubleSort(int *v, int n, long long *passos);
+void bubleSort(int *v, long n, long long *passos);
 
-void selectionSort(int *v, int n, long long *passos);
+void selectionSort(int *v, long n, long long *passos);
 
-void insertionSort(int *v, int n, long long *passos);
+void insertionSort(int *v, long n, long long *passos);
 
-int particiona(int *v, int inicio, int final, long long *passos);
+int particiona(int *v, int inicio, long final, long long *passos);
 
-void quickSort(int *v, int inicio, int fim, long long *passos);
+void quickSort(int *v, int inicio, long fim, long long *passos);
